refactor common methods in Page and FormXobject classes
* extracted all the resource methods (fonts, colorspaces, etc) out into
  a mixin
* made Page#resources and FormXobject private methods
