+++ lib/pdf/reader/form_xobject.rb
@@ -11,23 +11,12 @@ module PDF
     # This behaves and looks much like a limited PDF::Reader::Page class.
     #
     class FormXObject
+      include ResourceMethods
 
       def initialize(page, xobject)
         @page    = page
         @objects = page.objects
         @xobject = @objects.deref(xobject)
-      end
-
-      # Returns the resources that accompany this form.
-      #
-      def resources
-        @resources ||= @objects.deref(@xobject.hash[:Resources]) || {}
-      end
-
-      # Returns a Hash of XObjects that are available to this page
-      #
-      def xobjects
-        @objects.deref!(@resources[:XObject]) || {}
       end
 
       # return a hash of fonts used on this form.
@@ -37,7 +26,7 @@ module PDF
       # The values are a PDF::Reader::Font instances that provide access
       # to most available metrics for each font.
       #
-      def fonts
+      def font_objects
         raw_fonts = @objects.deref(resources[:Font] || {})
         ::Hash[raw_fonts.map { |label, font|
           [label, PDF::Reader::Font.new(@objects, @objects.deref(font))]
@@ -61,6 +50,12 @@ module PDF
       end
 
       private
+
+      # Returns the resources that accompany this form.
+      #
+      def resources
+        @resources ||= @objects.deref(@xobject.hash[:Resources]) || {}
+      end
 
       def callback(receivers, name, params=[])
         receivers.each do |receiver|
