+++ lib/pdf/reader/page.rb
@@ -12,6 +12,7 @@ module PDF
     # objects accessor to help walk the page dictionary in any useful way.
     #
     class Page
+      include ResourceMethods
 
       # lowlevel hash-like access to all objects in the underlying PDF
       attr_reader :objects
@@ -54,62 +55,6 @@ module PDF
             hash.merge!(@objects.deref(obj))
           end
         }
-      end
-
-      # Returns the resources that accompany this page. Includes
-      # resources inherited from parents.
-      #
-      def resources
-        @resources ||= @objects.deref(attributes[:Resources]) || {}
-      end
-
-      # Returns a Hash of color spaces that are available to this page
-      #
-      def color_spaces
-        @objects.deref!(resources[:ColorSpace]) || {}
-      end
-
-      # Returns a Hash of fonts that are available to this page
-      #
-      def fonts
-        @objects.deref!(resources[:Font]) || {}
-      end
-
-      # Returns a Hash of external graphic states that are available to this
-      # page
-      #
-      def graphic_states
-        @objects.deref!(resources[:ExtGState]) || {}
-      end
-
-      # Returns a Hash of patterns that are available to this page
-      #
-      def patterns
-        @objects.deref!(resources[:Pattern]) || {}
-      end
-
-      # Returns an Array of procedure sets that are available to this page
-      #
-      def procedure_sets
-        @objects.deref!(resources[:ProcSet]) || []
-      end
-
-      # Returns a Hash of properties sets that are available to this page
-      #
-      def properties
-        @objects.deref!(resources[:Properties]) || {}
-      end
-
-      # Returns a Hash of shadings that are available to this page
-      #
-      def shadings
-        @objects.deref!(resources[:Shading]) || {}
-      end
-
-      # Returns a Hash of XObjects that are available to this page
-      #
-      def xobjects
-        @objects.deref!(resources[:XObject]) || {}
       end
 
       # returns the plain text content of this page encoded as UTF-8. Any
@@ -166,6 +111,13 @@ module PDF
 
       def root
         root ||= objects.deref(@objects.trailer[:Root])
+      end
+
+      # Returns the resources that accompany this page. Includes
+      # resources inherited from parents.
+      #
+      def resources
+        @resources ||= @objects.deref(attributes[:Resources]) || {}
       end
 
       def content_stream(receivers, instructions)
